+++ ElasticApmAgent.sln
@@ -116,7 +116,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KafkaSample", "test\profile
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RabbitMqSample", "test\profiler\applications\RabbitMqSample\RabbitMqSample.csproj", "{1D6B0C67-42C8-4AB4-A795-B6FF8EF7196E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenTelemetrySample", "sample\OpenTelemetrySample\OpenTelemetrySample.csproj", "{3D31589A-CA44-4D6D-9944-26CE9640A283}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenTelemetrySample", "test\opentelemetry\OpenTelemetrySample\OpenTelemetrySample.csproj", "{3D31589A-CA44-4D6D-9944-26CE9640A283}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SatelliteAssemblySample", "test\profiler\applications\SatelliteAssemblySample\SatelliteAssemblySample.csproj", "{2B62E740-A493-42D6-BE5E-709BD07F877E}"
 EndProject
@@ -214,6 +214,10 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "startuphook", "startuphook", "{93686A26-AD8A-498C-BC0E-FF8258F3D20C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "applications", "applications", "{59F3FB6E-4B48-4E87-AF3B-78DFED427EF1}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "opentelemetry", "opentelemetry", "{1CDE0A67-0245-4CB0-A1B9-D17CCC388E82}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Elastic.Apm.OpenTelemetry.Tests", "test\opentelemetry\Elastic.Apm.OpenTelemetry.Tests\Elastic.Apm.OpenTelemetry.Tests.csproj", "{8EF0BEC9-C979-42B9-9447-F03570CCAA24}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -543,6 +547,10 @@ Global
 		{69C34D4F-13A5-43DE-A126-954B85F8A33E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{69C34D4F-13A5-43DE-A126-954B85F8A33E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{69C34D4F-13A5-43DE-A126-954B85F8A33E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8EF0BEC9-C979-42B9-9447-F03570CCAA24}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8EF0BEC9-C979-42B9-9447-F03570CCAA24}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8EF0BEC9-C979-42B9-9447-F03570CCAA24}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8EF0BEC9-C979-42B9-9447-F03570CCAA24}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -561,7 +569,6 @@ Global
 		{5D076C7F-1F8B-4B11-9910-48717D133963} = {3734A52F-2222-454B-BF58-1BA5C1F29D77}
 		{43F9247D-544B-49FB-9E50-FC236D90DE1A} = {267A241E-571F-458F-B04C-B6C4DE79E735}
 		{27563B4E-ECB1-4F1B-B9F1-22C2C165B270} = {3C791D9C-6F19-4F46-B367-2EC0F818762D}
-		{3D31589A-CA44-4D6D-9944-26CE9640A283} = {3C791D9C-6F19-4F46-B367-2EC0F818762D}
 		{BCE93684-0A40-4F5F-8FB5-94DE94B111E1} = {3C791D9C-6F19-4F46-B367-2EC0F818762D}
 		{C6A7CA94-9D06-4CA1-BAEA-43B607BF939D} = {3734A52F-2222-454B-BF58-1BA5C1F29D77}
 		{42F4F13C-0515-4501-A27B-9815BFF63807} = {C6A7CA94-9D06-4CA1-BAEA-43B607BF939D}
@@ -640,6 +647,9 @@ Global
 		{319E2094-D2E1-46C5-8985-9693CD8DDCBD} = {59F3FB6E-4B48-4E87-AF3B-78DFED427EF1}
 		{0ADF687D-4DB7-417E-80B4-27B3FD86875E} = {59F3FB6E-4B48-4E87-AF3B-78DFED427EF1}
 		{00AD0047-92A6-48CC-B67D-A3DB2812CE28} = {59F3FB6E-4B48-4E87-AF3B-78DFED427EF1}
+		{1CDE0A67-0245-4CB0-A1B9-D17CCC388E82} = {267A241E-571F-458F-B04C-B6C4DE79E735}
+		{3D31589A-CA44-4D6D-9944-26CE9640A283} = {1CDE0A67-0245-4CB0-A1B9-D17CCC388E82}
+		{8EF0BEC9-C979-42B9-9447-F03570CCAA24} = {1CDE0A67-0245-4CB0-A1B9-D17CCC388E82}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {69E02FD9-C9DE-412C-AB6B-5B8BECC6BFA5}
