+++ test/opentelemetry/OpenTelemetrySample/OpenTelemetrySample.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\Elastic.Apm\Elastic.Apm.csproj" />
+      <ProjectReference Include="$(SolutionRoot)\src\Elastic.Apm\Elastic.Apm.csproj" />
     </ItemGroup>
 
 </Project>
